move spill pool testing for aggregate functions to physical-plan/src/aggregates
